move to module-based struct, rename imports
