+++ garak/__init__.py
@@ -1,0 +1,4 @@
+"""Top-level package for garak"""
+
+__app_name__ = "garak"
+__version__ = "0.9"
