+++ garak/__main__.py
@@ -1,0 +1,9 @@
+"""garak entry point script"""
+
+from garak import cli, __app_name__
+
+def main():
+    cli.main()
+
+if __name__ == "__main__":
+    main()
