+++ garak/detectors/knownbadsignatures.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python3
 
-from detectors.base import StringDetector
-import probes.knownbadsignatures
+from garak.detectors.base import StringDetector
+from garak.probes import knownbadsignatures
 
 
 class EICAR(StringDetector):
     def __init__(self):
         self.name = "eicar"
-        self.substrings = [probes.knownbadsignatures.eicar_signature]
+        self.substrings = [knownbadsignatures.eicar_signature]
         super().__init__(self.substrings)
         self.description = "Is the EICAR test string present?"
         self.matchtype = "str"
@@ -16,7 +16,7 @@ class EICAR(StringDetector):
 class GTUBE(StringDetector):
     def __init__(self):
         self.name = "gtube"
-        self.substrings = [probes.knownbadsignatures.gtube_signature]
+        self.substrings = [knownbadsignatures.gtube_signature]
         super().__init__(self.substrings)
         self.description = "Is the GTUBE test string present?"
         self.matchtype = "str"
@@ -25,7 +25,7 @@ class GTUBE(StringDetector):
 class GTphish(StringDetector):
     def __init__(self):
         self.name = "gtphish"
-        self.substrings = [probes.knownbadsignatures.gtphish_signature]
+        self.substrings = [knownbadsignatures.gtphish_signature]
         super().__init__(self.substrings)
         self.description = "Is the GTphish test string present?"
         self.matchtype = "str"
